### 2023-07-12 #5652 @editorialbot
[PRE REVIEW]: MicroTracker.jl: A Julia package for microbot research
**Submitting author:** @czimm79 (<a href="http://orcid.org/0000-0002-0863-0723">Coy Zimmermann</a>)
**Repository:** https://github.com/czimm79/MicroTracker.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.2
