### 2023-08-01 #5652 @Kevin-Mattheus-Moerman
@beniroquai, @RainerHeintzmann, @roflmaostc, @mkitti, @BioTurboNick, @tlnagy would you be interested in reviewing this submission for [JOSS (the Journal of Open Source Software)](https://joss.theoj.org)? The work is entitled "MicroTracker.jl: A Julia package for microbot research", and features the use of Julia to track features in Microscopy data.

JOSS reviews are conduced in the open here on GitHub, and focus on evaluating [the software](https://github.com/czimm79/MicroTracker.jl), and reviewing [a short paper](https://raw.githubusercontent.com/openjournals/joss-papers/joss.05652/joss.05652/10.21105.joss.05652.pdf).

You can let me know here in a comment if you are interested in reviewing. I'll then add you as a reviewer and a dedicated review issue will be opened once enough reviewers are attracted.  Here is an example of an ongoing review for a different submission, which shows how reviewers are guided through the process by a set of checkboxes: https://github.com/openjournals/joss-reviews/issues/5549.
You can also check our [review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) for more information on how reviews for JOSS are conducted.

Thanks!
