### 2023-08-01 #5652 @beniroquai
Looks great, but unfortunately I'm not enough of a Julia expert to judge about the code. @RainerHeintzmann, are you keen to review this manuscript? :)
