### 2023-09-04 #5804 @roflmaostc
Well written and extensive documentation!

However, I would recommend to upload the data elsewhere instead of uploading ~300MB of images to GitHub.
Also, running the default tutorial, it was not quite clear where the data is coming from. Is there a default path?
My first run was failing because of path issues and couldn't find out how to fix it.
