### 2023-09-05 #5804 @czimm79
Hi @roflmaostc, thanks for volunteering to review our submission!

After looking into potential alternatives for hosting the example images, it seems like the best practice for Julia packages is to use the [Artifacts](https://pkgdocs.julialang.org/v1/artifacts/) feature of Pkg.jl. I will get to work on this and create a new fork of MicroTracker which uses Artifacts.

Currently, MicroTracker is designed so the user's working directory is the project folder. This allows the microscopy video, segmenting results, and linking results to stay organized and referenceable throughout the pipeline. This project folder is created using the [`create_project_here`](https://czimm79.github.io/MicroTracker.jl/dev/api/#MicroTracker.create_project_here-Tuple{}) function. Subsequent functions like [`batch_particle_data_to_linked_data`](https://czimm79.github.io/MicroTracker.jl/dev/api/#MicroTracker.batch_particle_data_to_linked_data-Tuple{Dict,%20NamedTuple}) reference the folders inside the project folder to load and save data. We will work on adding some path checking error handing and documentation to clarify the functions that need to be run from the project working directory to try and mitigate path issues.
