### 2023-10-04 #5804 @mkitti
Installation goes quite smoothly. The package also has some Python package dependencies including `numpy`, `pandas`, and `trackpy`. It would be nice to more clearly document this in the Setup part of the documentation with a note that this will install these packages in the default ROOT conda environment. It would be even better if it could better documented how to use a bespoke conda environment for this purpose. I would be happy to work with the authors on developing that documentation since I am involved with PyCall.jl and Conda.jl.

Additionally, for the purpose of the paper, could the authors also document the versions of all dependencies used. For Julia dependencies this could be in the form of a (Julia)Manifest.toml. While this Manifest.toml should not necessarily be at the package root, perhaps it could be included in the `paper` directory. If a Project.toml is included along with this manifest, then the project could be activated and instantiated. To clarify, the purpose of this pair of Project.toml and Manifest.toml is needed to document the project used to generate the examples in the paper. Additonally, the versions of the Python packages should be documented. This could be done by exporting an environment.yml file via `conda`. Establishing the versions of all dependencies is required for future reproducibility.
