### 2023-10-04 #5804 @mkitti
# Community Guidelines

It is documented how to contribute to the software. However documentation on these issues are currently absent.
* Report issues or problems with the software
* Seek support

Generally for Julia projects like this `Github issues` may be used and/or one of the [Julia community channels](https://julialang.org/community/) such as Slack or Discourse. The authors should explicitly note how they prefer to be contacted, how to report issues, and how to seek support.
