### 2023-10-04 #5804 @mkitti
# Automated testing

I am experiencing a test failure on my Windows machine. I will investigate further. My guess is that this is related to the Python environment issues mentioned above.

```
julia> versioninfo()
Julia Version 1.9.3
Commit bed2cd540a (2023-08-24 14:43 UTC)
Build Info:
  Official https://julialang.org/ release
Platform Info:
  OS: Windows (x86_64-w64-mingw32)
  CPU: 48 × Intel(R) Xeon(R) Gold 5220R CPU @ 2.20GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-14.0.6 (ORCJIT, cascadelake)
  Threads: 1 on 96 virtual cores
```

<summary>Test failure in python_interactions.jl</summary>

```julia
Python functions: Error During Test at C:\Users\kittisopikulm\.julia\dev\MicroTracker\test\python_interactions.jl:10
  Got exception outside of a @test
  PyError ($(Expr(:escape, :(ccall(#= C:\Users\kittisopikulm\.julia\packages\PyCall\ilqDX\src\pyfncall.jl:43 =# @pysym(:PyObject_Call), PyPtr, (PyPtr, PyPtr, PyPtr), o, pyargsptr, kw))))) <class 'ValueError'>
  ValueError("Invalid file path or buffer object type: <class 'PyIO'>")
    File "C:\Users\kittisopikulm\.julia\conda\3\lib\site-packages\pandas\util\_decorators.py", line 311, in wrapper
      return func(*args, **kwargs)
    File "C:\Users\kittisopikulm\.julia\conda\3\lib\site-packages\pandas\io\parsers\readers.py", line 586, in read_csv
      return _read(filepath_or_buffer, kwds)
    File "C:\Users\kittisopikulm\.julia\conda\3\lib\site-packages\pandas\io\parsers\readers.py", line 482, in _read
      parser = TextFileReader(filepath_or_buffer, **kwds)
    File "C:\Users\kittisopikulm\.julia\conda\3\lib\site-packages\pandas\io\parsers\readers.py", line 811, in __init__
      self._engine = self._make_engine(self.engine)
    File "C:\Users\kittisopikulm\.julia\conda\3\lib\site-packages\pandas\io\parsers\readers.py", line 1040, in _make_engine
      return mapping[engine](self.f, **self.options)  # type: ignore[call-arg]
    File "C:\Users\kittisopikulm\.julia\conda\3\lib\site-packages\pandas\io\parsers\c_parser_wrapper.py", line 51, in __init__
      self._open_handles(src, kwds)
    File "C:\Users\kittisopikulm\.julia\conda\3\lib\site-packages\pandas\io\parsers\base_parser.py", line 222, in _open_handles
      self.handles = get_handle(
    File "C:\Users\kittisopikulm\.julia\conda\3\lib\site-packages\pandas\io\common.py", line 609, in get_handle
      ioargs = _get_filepath_or_buffer(
    File "C:\Users\kittisopikulm\.julia\conda\3\lib\site-packages\pandas\io\common.py", line 396, in _get_filepath_or_buffer
      raise ValueError(msg)

  Stacktrace:
    [1] pyerr_check
      @ C:\Users\kittisopikulm\.julia\packages\PyCall\ilqDX\src\exception.jl:75 [inlined]
    [2] pyerr_check
      @ C:\Users\kittisopikulm\.julia\packages\PyCall\ilqDX\src\exception.jl:79 [inlined]
    [3] _handle_error(msg::String)
      @ PyCall C:\Users\kittisopikulm\.julia\packages\PyCall\ilqDX\src\exception.jl:96
    [4] macro expansion
      @ C:\Users\kittisopikulm\.julia\packages\PyCall\ilqDX\src\exception.jl:110 [inlined]
    [5] #107
      @ C:\Users\kittisopikulm\.julia\packages\PyCall\ilqDX\src\pyfncall.jl:43 [inlined]
    [6] disable_sigint
      @ .\c.jl:473 [inlined]
    [7] __pycall!
      @ C:\Users\kittisopikulm\.julia\packages\PyCall\ilqDX\src\pyfncall.jl:42 [inlined]
    [8] _pycall!(ret::PyCall.PyObject, o::PyCall.PyObject, args::Tuple{IOBuffer}, nargs::Int64, kw::Ptr{Nothing})
      @ PyCall C:\Users\kittisopikulm\.julia\packages\PyCall\ilqDX\src\pyfncall.jl:29
    [9] _pycall!(ret::PyCall.PyObject, o::PyCall.PyObject, args::Tuple{IOBuffer}, kwargs::Base.Pairs{Symbol, Union{}, Tuple{}, NamedTuple{(), Tuple{}}})
      @ PyCall C:\Users\kittisopikulm\.julia\packages\PyCall\ilqDX\src\pyfncall.jl:11
   [10] (::PyCall.PyObject)(args::IOBuffer; kwargs::Base.Pairs{Symbol, Union{}, Tuple{}, NamedTuple{(), Tuple{}}})
      @ PyCall C:\Users\kittisopikulm\.julia\packages\PyCall\ilqDX\src\pyfncall.jl:86
   [11] (::PyCall.PyObject)(args::IOBuffer)
      @ PyCall C:\Users\kittisopikulm\.julia\packages\PyCall\ilqDX\src\pyfncall.jl:86
   [12] jldf_to_pydf(jldf::DataFrame)
      @ MicroTracker C:\Users\kittisopikulm\.julia\dev\MicroTracker\src\developer_utilities.jl:55
   [13] macro expansion
      @ C:\Users\kittisopikulm\.julia\dev\MicroTracker\test\python_interactions.jl:25 [inlined]
   [14] macro expansion
      @ C:\Users\kittisopikulm\.julia\juliaup\julia-1.9.3+0.x64.w64.mingw32\share\julia\stdlib\v1.9\Test\src\Test.jl:1498 [inlined]
   [15] top-level scope
      @ C:\Users\kittisopikulm\.julia\dev\MicroTracker\test\python_interactions.jl:11
   [16] include(fname::String)
      @ Base.MainInclude .\client.jl:478
   [17] top-level scope
      @ C:\Users\kittisopikulm\.julia\dev\MicroTracker\test\runtests.jl:54
   [18] include(fname::String)
      @ Base.MainInclude .\client.jl:478
   [19] top-level scope
      @ none:6
   [20] eval
      @ .\boot.jl:370 [inlined]
   [21] exec_options(opts::Base.JLOptions)
      @ Base .\client.jl:280
   [22] _start()
      @ Base .\client.jl:522
Test Summary:    | Pass  Error  Total  Time
Python functions |    4      1      5  9.5s
ERROR: LoadError: Some tests did not pass: 4 passed, 0 failed, 1 errored, 0 broken.
in expression starting at C:\Users\kittisopikulm\.julia\dev\MicroTracker\test\python_interactions.jl:10
in expression starting at C:\Users\kittisopikulm\.julia\dev\MicroTracker\test\runtests.jl:54
ERROR: Package MicroTracker errored during testing
```

Also, I'm noting that the automated Github Actions testing does pass. However, there is [some output](https://github.com/czimm79/MicroTracker.jl/actions/runs/5502626073/job/14895753095) that should be investigated concerning the plot tests.

```
qt.qpa.xcb: could not connect to display
qt.qpa.plugin: Could not load the Qt platform plugin "xcb" in "" even though it was found.
This application failed to start because no Qt platform plugin could be initialized. Reinstalling the application may fix this problem.

Available platform plugins are: linuxfb, minimal, offscreen, vnc, xcb.

Aborted (core dumped)
connect: Connection refused
GKS: can't connect to GKS socket application

GKS: Open failed in routine OPEN_WS
GKS: GKS not in proper state. GKS must be either in the state WSOP or WSAC in routine ACTIVATE_WS
qt.qpa.xcb: could not connect to display
qt.qpa.plugin: Could not load the Qt platform plugin "xcb" in "" even though it was found.
This application failed to start because no Qt platform plugin could be initialized. Reinstalling the application may fix this problem.

Available platform plugins are: linuxfb, minimal, offscreen, vnc, xcb.

Aborted (core dumped)
connect: Connection refused
GKS: can't connect to GKS socket application

GKS: Open failed in routine OPEN_WS
GKS: GKS not in proper state. GKS must be either in the state WSOP or WSAC in routine ACTIVATE_WS
```
