### 2023-10-04 #5804 @mkitti
# State of the Field

The paper mentions several other tracking packages written for other languages including Python and Java. It also happens to be able to integrate with some of them.

There is another Julia tracking package called [BlobTracking.jl](https://github.com/baggepinnen/BlobTracking.jl). Please compare and differentiate this package from BlobTracking.jl.
