### 2023-10-05 #5804 @mkitti
Current review summary:

MicroTracker.jl is well packaged and documented. The package is used to track microbots, a form of active matter inspired by natural phenomena. While bespoke for a particular materials science experimental design, the package seems to have potential for general tracking problems, greatly expanding the number of potential users.

I appreciate that the authors have taken the time to mention some variety of other tracking packages. However, the tracking field is quite large with numerous tracking packages available. Because of this, I feel the authors have understated the "State of the Field". In particular several reviews cover the results of the [particle tracking](https://www.nature.com/articles/nmeth.2808) and [cell tracking](https://www.nature.com/articles/s41592-023-01879-y) challenges. Mentioning the existence of these challenges would be helpful in expanding the discussion of the field.

In the expanded context, the authors should motivate further the creation of their software and in particular describe what makes microbots different from particles or cells. The authors currently that "MicroTracker.jl allows for size, shape, and rotation rate tracking of microbots which regularly enter and leave the video frame". However, it is unclear to me how that differs from other tracking problems that appear in the previously mentioned challenges.

What would help is to first explain the nature of the problem that this package is meant to address. In particular, what are the unique challenges that tracking microbots present? What are the unique or particular solutions that this package implements? Some hint of this is given in the paragraph starting on line 48, but this needs to be expanded and probably should some much earlier in the paper.

There are a few technical points that I've outlined in detail above.

This Julia package has Python package dependencies. There is no mention of this in the Quick Start or Setup portions of the documentation. However, the interface between Python and Julia often needs careful configuration. In particular, the authors use the function `pyimport_conda` from PyCall.jl. Proper use of this function requires that both the packages Conda.jl and PyCall.jl be consistently configured. The authors should at minimum note the use of these dependencies and reference the documentation for those packages.

The versions of the Julia and Python dependencies should be listed for reproducibility.

Automated testing does exist in the form of Julia tests and is run by Github Actions on each commit. This obviously only works with a properly installed package including with the Python dependencies.

Additionally, the community guidelines requirement of this journal are not fully met as per the above.

Overall, MicroTracker.jl is a very promising package in a fastastic state. Some improvements could be made. In particular, explaining the problem being solved by this package thoroughly will help potential users evaluate if they have similar problems that could be solved either by this package directly or could be solved by studing the open source code of this package. Some technical improvements could be made or documented to help users forestall potential issues. I believe these issues could be addressed by the authors in a reasonable amount of time, and I look forward to seeing those changes.
