### 2023-10-05 #5804 @mkitti
Some technical challenges I had to overcome to properly configure the Julia - Python interface:
1. PyCall.jl initially detected the system `python3` installed in Ubuntu rather than the `python` executable installed by conda in Conda.jl. To properly configure this, I needed to either remove `python3` from my executable path or configure my executable path to be of the conda environment followed by using `Pkg.build` to rebuild PyCall.jl.
2. Setting the `$PYTHON` environment variable alone was not compatible with the author's use of `pyimport_conda` since in this mode PyCall.jl is not aware of the conda installation. In this case, `Conda.add` needed to be invoked manually before `pyimport_conda` could succeed.

I believe the author's current installation process may succeed in the conditions where there is no `python` executable on the current path, [PyCall.jl uses Python from the Conda.jl conda install](https://github.com/JuliaPy/PyCall.jl/blob/2a9f07725b1bce8c07ac5437d66701dd9579a221/deps/build.jl#L67), and then configures itself to use both the executables `python` and `conda` from the same environment.

It is not the author's responsibility to solve and lessen this complex process, but it would helpful to note that this is part of the installation process and where to get help if the above process goes awry as it did for me on my Windows and Linux machines. Fortunately, I have some expertise in this matter and was able to resolve them. It would be helpful if the author's could point other potential users to the PyCall.jl and Conda.jl documentation for troubleshooting.
