### 2023-10-10 czimm79/MicroTracker.jl#40 @BioTurboNick
Unable to install - missing python packages
```
[ Info: Installing trackpy via the Conda trackpy package...
[ Info: Running `conda install -y trackpy` in root environment
Collecting package metadata (current_repodata.json): done
Solving environment: failed with initial frozen solve. Retrying with flexible solve.
Solving environment: failed with initial frozen solve. Retrying with flexible solve.

PackagesNotFoundError: The following packages are not available from current channels:

  - trackpy

Current channels:

  - https://repo.anaconda.com/pkgs/main/win-64
  - https://repo.anaconda.com/pkgs/main/noarch
  - https://repo.anaconda.com/pkgs/r/win-64
  - https://repo.anaconda.com/pkgs/r/noarch
  - https://repo.anaconda.com/pkgs/msys2/win-64
  - https://repo.anaconda.com/pkgs/msys2/noarch

To search for alternate channels that may provide the conda package you're
looking for, navigate to

    https://anaconda.org

and use the search bar at the top of the page.

ERROR: InitError: failed process: Process(setenv(`'C:\Users\nicho\.julia\conda\3\Scripts\conda.exe' install -y trackpy`,["WINDIR=C:\\WINDOWS", "PATH=C:\\Users\\nicho\\.julia\\conda\\3\\Library\\bin;C:\\Users\\nicho\\.julia\\conda\\3\\Library\\bin;C:\\Program Files\\WindowsApps\\Microsoft.PowerShell_7.3.7.0_x64__8wekyb3d8bbwe;C:\\Program Files (x86)\\NVIDIA Corporation\\PhysX\\Common;C:\\WINDOWS\\system32;C:\\WINDOWS;C:\\WINDOWS\\System32\\Wbem;C:\\WINDOWS\\System32\\WindowsPowerShell\\v1.0\\;C:\\WINDOWS\\System32\\OpenSSH\\;C:\\Program Files\\Amazon\\AWSCLIV2\\;C:\\Program Files\\Microsoft SQL Server\\Client SDK\\ODBC\\110\\Tools\\Binn\\;C:\\Program Files (x86)\\Microsoft SQL Server\\120\\Tools\\Binn\\;C:\\Program Files\\Microsoft SQL Server\\120\\Tools\\Binn\\;C:\\Program Files\\Microsoft SQL Server\\120\\DTS\\Binn\\;C:\\Program Files (x86)\\Windows Kits\\8.1\\Windows Performance Toolkit\\;C:\\Program Files\\Git\\cmd;C:\\Program Files\\dotnet\\;C:\\WINDOWS\\system32;C:\\WINDOWS;C:\\WINDOWS\\System32\\Wbem;C:\\WINDOWS\\System32\\WindowsPowerShell\\v1.0\\;C:\\WINDOWS\\System32\\OpenSSH\\;C:\\Program Files\\Docker\\Docker\\resources\\bin;C:\\Program Files\\PowerShell\\7\\;C:\\Users\\nicho\\AppData\\Local\\Microsoft\\WindowsApps;C:\\Users\\nicho\\AppData\\Local\\gitkraken\\bin;C:\\Users\\nicho\\AppData\\Local\\Programs\\Microsoft VS Code\\bin;C:\\ProgramData\\nicho\\GitHubDesktop\\bin;C:\\Users\\nicho\\AppData\\Local\\Programs\\MiKTeX\\miktex\\bin\\x64\\;C:\\Users\\nicho\\AppData\\Local\\Microsoft\\WindowsApps;C:\\Users\\nicho\\.dotnet\\tools;", "USERDOMAIN_ROAMINGPROFILE=AURORASURFACE8", "WSLENV=WT_SESSION:WT_PROFILE_ID:", "LOCALAPPDATA=C:\\Users\\nicho\\AppData\\Local", "HOMEPATH=\\Users\\nicho", "PROCESSOR_IDENTIFIER=Intel64 Family 6 Model 140 Stepping 1, GenuineIntel", "NUMBER_OF_PROCESSORS=8", "PATHEXT=.COM;.EXE;.BAT;.CMD;.VBS;.VBE;.JS;.JSE;.WSF;.WSH;.MSC;.CPL", "SESSIONNAME=Console"  …  "PROCESSOR_LEVEL=6", "SW_SIM_HYDRA=C:\\Program Files\\Common Files\\SolidWorks Shared\\Simulation Worker Agent\\", "SYSTEMDRIVE=C:", "=C:=C:\\Users\\nicho", "PROGRAMW6432=C:\\Program Files", "TEMP=C:\\Users\\nicho\\AppData\\Local\\Temp", "HOMEDRIVE=C:", "OPENBLAS_MAIN_FREE=1", "PROCESSOR_ARCHITECTURE=AMD64", "PYTHONIOENCODING=UTF-8"]), ProcessExited(1)) [1]

Stacktrace:
  [1] pipeline_error
    @ .\process.jl:565 [inlined]
  [2] run(::Cmd; wait::Bool)
    @ Base .\process.jl:480
  [3] run
    @ .\process.jl:477 [inlined]
  [4] runconda(args::Cmd, env::String)
    @ Conda C:\Users\nicho\.julia\packages\Conda\2lg2O\src\Conda.jl:128
  [5] add(pkg::String, env::String; channel::String, satisfied_skip_solve::Bool, args::Cmd)
    @ Conda C:\Users\nicho\.julia\packages\Conda\2lg2O\src\Conda.jl:256
  [6] add (repeats 2 times)
    @ C:\Users\nicho\.julia\packages\Conda\2lg2O\src\Conda.jl:239 [inlined]
  [7] pyimport_conda
    @ C:\Users\nicho\.julia\packages\PyCall\ilqDX\src\PyCall.jl:721
  [8] pyimport_conda
    @ C:\Users\nicho\.julia\packages\PyCall\ilqDX\src\PyCall.jl:715 [inlined]
  [9] __init__()
    @ MicroTracker C:\Users\nicho\.julia\packages\MicroTracker\2n7hT\src\MicroTracker.jl:16
 [10] register_restored_modules(sv::Core.SimpleVector, pkg::Base.PkgId, path::String)
    @ Base .\loading.jl:1115
 [11] _include_from_serialized(pkg::Base.PkgId, path::String, ocachepath::String, depmods::Vector{Any})
    @ Base .\loading.jl:1061
 [12] _require_search_from_serialized(pkg::Base.PkgId, sourcepath::String, build_id::UInt128)
    @ Base .\loading.jl:1506
 [13] _require(pkg::Base.PkgId, env::String)
    @ Base .\loading.jl:1783
 [14] _require_prelocked(uuidkey::Base.PkgId, env::String)
    @ Base .\loading.jl:1660
 [15] macro expansion
    @ .\loading.jl:1648 [inlined]
 [16] macro expansion
    @ .\lock.jl:267 [inlined]
 [17] require(into::Module, mod::Symbol)
    @ Base .\loading.jl:1611
during initialization of module MicroTracker

caused by: PyError (PyImport_ImportModule

The Python package trackpy could not be imported by pyimport. Usually this means
that you did not install trackpy in the Python version being used by PyCall.

PyCall is currently configured to use the Julia-specific Python distribution
installed by the Conda.jl package.  To install the trackpy module, you can
use `pyimport_conda("trackpy", PKG)`, where PKG is the Anaconda
package that contains the module trackpy, or alternatively you can use the
Conda package directly (via `using Conda` followed by `Conda.add` etcetera).

Alternatively, if you want to use a different Python distribution on your
system, such as a system-wide Python (as opposed to the Julia-specific Python),
you can re-configure PyCall with that Python.   As explained in the PyCall
documentation, set ENV["PYTHON"] to the path/name of the python executable
you want to use, run Pkg.build("PyCall"), and re-launch Julia.

) <class 'ModuleNotFoundError'>
ModuleNotFoundError("No module named 'trackpy'")

Stacktrace:
  [1] pyimport(name::String)
    @ PyCall C:\Users\nicho\.julia\packages\PyCall\ilqDX\src\PyCall.jl:558
  [2] pyimport_conda
    @ C:\Users\nicho\.julia\packages\PyCall\ilqDX\src\PyCall.jl:716
  [3] pyimport_conda
    @ C:\Users\nicho\.julia\packages\PyCall\ilqDX\src\PyCall.jl:715 [inlined]
  [4] __init__()
    @ MicroTracker C:\Users\nicho\.julia\packages\MicroTracker\2n7hT\src\MicroTracker.jl:16
  [5] register_restored_modules(sv::Core.SimpleVector, pkg::Base.PkgId, path::String)
    @ Base .\loading.jl:1115
  [6] _include_from_serialized(pkg::Base.PkgId, path::String, ocachepath::String, depmods::Vector{Any})
    @ Base .\loading.jl:1061
  [7] _require_search_from_serialized(pkg::Base.PkgId, sourcepath::String, build_id::UInt128)
    @ Base .\loading.jl:1506
  [8] _require(pkg::Base.PkgId, env::String)
    @ Base .\loading.jl:1783
  [9] _require_prelocked(uuidkey::Base.PkgId, env::String)
    @ Base .\loading.jl:1660
 [10] macro expansion
    @ .\loading.jl:1648 [inlined]
 [11] macro expansion
    @ .\lock.jl:267 [inlined]
 [12] require(into::Module, mod::Symbol)
    @ Base .\loading.jl:1611
```

```
Julia Version 1.9.3
Commit bed2cd540a (2023-08-24 14:43 UTC)
Build Info:
  Official https://julialang.org/ release
Platform Info:
  OS: Windows (x86_64-w64-mingw32)
  CPU: 8 × 11th Gen Intel(R) Core(TM) i5-1135G7 @ 2.40GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-14.0.6 (ORCJIT, tigerlake)
  Threads: 1 on 8 virtual cores
Environment:
  JULIA_IMAGE_THREADS = 1
  JULIA_PKG_USE_CLI_GIT = true
```
