### 2023-10-10 #5804 @BioTurboNick
I was unable to install the code as directed here (https://czimm79.github.io/MicroTracker.jl/dev/quickstart/), thanks to a Python/Conda issue, documented here: https://github.com/czimm79/MicroTracker.jl/issues/40
