### 2023-10-10 czimm79/MicroTracker.jl#40 @BioTurboNick
I did get it to install as expected on a different machine. However, some consideration of Julia/Python connection issues should probably be noted and addressed for naive users.
```
Julia Version 1.9.3
Commit bed2cd540a (2023-08-24 14:43 UTC)
Build Info:
  Official https://julialang.org/ release
Platform Info:
  OS: Windows (x86_64-w64-mingw32)
  CPU: 12 × Intel(R) Core(TM) i7-8700 CPU @ 3.20GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-14.0.6 (ORCJIT, skylake)
  Threads: 1 on 12 virtual cores
Environment:
  JULIA_IMAGE_THREADS = 1
``
