### 2023-10-11 czimm79/MicroTracker.jl#41 @BioTurboNick
Paper - summary incomplete, contains jargon
The summary focuses on introducing the field and the problem the paper/software is solving. However, what it doesn't do is summarize how the software is solving the problem. The summary should be enough to communicate to someone what your software is doing and what is interesting about it if that's all they ever read about it.

I also recommend checking for jargon usage. In particular, the first sentence defines "microbots" in terms that may be unfamiliar to a nonspecialist, particularly "active matter" and "remote work". The following sentences give a sense of what they may be, but only because I'm a biologist familiar with bacterial flagella and the rolling process of immune cells. Other jargon: "actuation methods", "remote magnetic actuation".
