### 2023-10-11 czimm79/MicroTracker.jl#42 @BioTurboNick
Paper - Break up Statement of Need
I think the Statement of Need section has the needed information, but is overloaded. Roughly, paragraphs 1, 4, and 5 seem to be most relevant. The implementation details in paragraphs 2 and 3 could be moved to a following custom section so that the Statement of Need can be more concise in answering the questions it should answer.
