### 2023-10-11 #5804 @BioTurboNick
Noted that the Statement of Need seems to be overloaded, advised to move the implementation details into a different section: https://github.com/czimm79/MicroTracker.jl/issues/42
