### 2023-10-11 czimm79/MicroTracker.jl#43 @BioTurboNick
Documentation - Statement of need expanded
The README contains a short paragraph at the top that meets the statement of need, but perhaps could be fleshed out a bit more. That section should probably also appear on the documentation home page, for users that may not see the repository.
