### 2023-10-11 #5804 @BioTurboNick
I think the documentation's statement of need could be expanded placed in the formal docs, not just in the README: https://github.com/czimm79/MicroTracker.jl/issues/43
