### 2023-10-11 czimm79/MicroTracker.jl#41 @mkitti
A clear description of the problem and its distinct features would be useful.

1. It would help the reader understand what problem is being solved.
2. It would help the reader grasp the solutions in this software.
3. It would help the reader evaluate if this software might help them solve *their* problem.
