### 2023-11-08 a582785
general documentation changes
index links, docs statement of need expanded, statement of need copied to home page of docs, contribution section in readme
