+++ docs/src/index.md
@@ -4,7 +4,7 @@ CurrentModule = MicroTracker
 
 # Home
 
-Welcome to the documentation for [MicroTracker](https://github.com/czimm79/MicroTracker.jl) - a Julia package for analyzing microscopy videos of microbots.
+Welcome to the documentation for [MicroTracker.jl](https://github.com/czimm79/MicroTracker.jl)! This is a feature-complete [Julia](https://julialang.org/) package that tracks and analyzes moving [microbots](https://www.nature.com/articles/s41467-020-19322-7) in microscopy video. This package tracks critical microbot metrics including size, velocity, and rotation rate. Combined with easy batch processing of videos, MicroTracker enables high-throughput analysis across multiple global experimental variables. For more information on microbots and this package, please see the short [paper](https://github.com/czimm79/MicroTracker.jl/blob/master/paper/paper.md).
 
 ## Get Started
 Follow an abridged version of the manual with included sample data and videos on the [Quick Start](@ref) page. This page also outlines the layout of the documentation.
