+++ README.md
@@ -4,16 +4,14 @@
 [![Build Status](https://github.com/czimm79/MicroTracker.jl/actions/workflows/CI.yml/badge.svg?branch=master)](https://github.com/czimm79/MicroTracker.jl/actions/workflows/CI.yml?query=branch%3Amaster)
 [![Coverage](https://codecov.io/gh/czimm79/MicroTracker.jl/branch/master/graph/badge.svg)](https://codecov.io/gh/czimm79/MicroTracker.jl)
 
-MicroTracker.jl is a feature-complete [Julia](https://julialang.org/) package that tracks and analyzes moving [microbots](https://www.nature.com/articles/s41467-020-19322-7) in microscopy video. This package allows tracking of critical microbot metrics including size, velocity, and rotation rate, enabling high-throughput analysis of experimental data.
+MicroTracker.jl is a feature-complete [Julia](https://julialang.org/) package that tracks and analyzes moving [microbots](https://www.nature.com/articles/s41467-020-19322-7) in microscopy video. This package tracks critical microbot metrics including size, velocity, and rotation rate. Combined with easy batch processing of videos, MicroTracker enables high-throughput analysis across multiple global experimental variables. For more information on microbots and this package, please see the short [paper](https://github.com/czimm79/MicroTracker.jl/blob/master/paper/paper.md).
 
 https://github.com/czimm79/MicroTracker.jl/assets/49537407/12c9a035-8b00-4576-a21b-ec7b8fa87629
-
-For more information on microbots and this package, please see the short [paper](https://github.com/czimm79/MicroTracker.jl/blob/master/paper/paper.md).
 
 To get started, view the docs by clicking the [![Docs](https://img.shields.io/badge/docs-latest-blue.svg)](https://czimm79.github.io/MicroTracker.jl/dev/) badge and click **Quick Start** on the left side. This package is designed for users with limited coding or Julia experience. Please raise an issue or start a pull request if anything could be improved.
 
 ## Collaboration/Contributing
-MicroTracker is built for collaboration and use across the microbot research field. To get started contributing and improving the package, see the [setup guide](https://czimm79.github.io/MicroTracker.jl/dev/settingupdev/).
+MicroTracker is built for collaboration and use across the microbot research field. To get started contributing and improving the package, see the [setup guide](https://czimm79.github.io/MicroTracker.jl/dev/settingupdev/). To report an issue, use the Issues tab 👆. Lastly, to seek support or discuss MicroTracker, please use the Discussions tab.
 
 ## Extra information
 Various components of MicroTracker have been used in scholarly works[^1][^2][^3][^4]. This package open sources, combines, and tests the tools written for microbot tracking as part of Coy Zimmermann's PhD thesis work 2018-2023 on magnetically propelled microwheels. This work was performed in the [Marr Group](https://chemeng.mines.edu/project/marr-david/) at the Colorado School of Mines and with collaboration of the [Neeves Lab](https://neeveslab.com/) at the University of Colorado Denver, Anschutz Medical Campus. MicroTracker uses and builds on work from colloidal science researchers[^5][^6].
