### 2023-11-09 fc0f771
conda-forge and docs
first commit, verify adding conda-forge channel does not break tests
