+++ docs/src/quickstart.md
@@ -4,14 +4,14 @@ This page is an abridged version of the documentation which runs through using M
 
 - **Install Julia** from their [website](https://julialang.org/downloads/platform/), making sure to add julia to your PATH. ([Download and install Julia](@ref))
 - **Open a terminal** in a new empty folder, and type `julia` to open a julia REPL in that directory. ([Open a Julia REPL in a directory](@ref))
-- **Create a Julia environment and install MicroTracker** into it using the following two commands. You can paste these into the Julia REPL.
+- **Create a Julia environment and install MicroTracker** into it using the following two commands. You can paste these into the Julia REPL. 
 ```
 ] activate .
 ```
 ```
 add MicroTracker
 ```
-- **Import MicroTracker** into your namespace using the following command. Make sure you press backspace to get out of the [Pkg mode](https://docs.julialang.org/en/v1/stdlib/Pkg/) before running this command.
+- **Import MicroTracker** into your namespace using the following command. Make sure you press backspace to get out of the [Pkg mode](https://docs.julialang.org/en/v1/stdlib/Pkg/) before running this command. Note that MicroTracker also automatically installs and imports a couple Python functions, see [Python interface](@ref) if any issues arise.
 
 ```julia
 using MicroTracker
