### 2023-11-09 aba58b7
warn->info and clean test output
