+++ .github/workflows/CI.yml
@@ -22,6 +22,7 @@ jobs:
           - 'nightly'
         os:
           - ubuntu-latest
+          - windows-latest
         arch:
           - x64
     env:
