### 2023-11-09 7bd7992
disable showing plots during test
fix CI warnings when trying to plot in a headless environment
