+++ test/test_plots.jl
@@ -3,6 +3,8 @@ macro isplot(ex) # @isplot macro to streamline tests, from Plots.jl
 end
 
 @testset "plotting" begin
+    gr(show=false)  # supress showing plots 
+
     linked_data = cd(()->load_linked_data("(B_mT=(2, Float64), FPS=(3, Float64), f_Hz=(1, Int64)) - (MPP = 0.605, SEARCH_RANGE_MICRONS = 1000, MEMORY = 0, STUBS_SECONDS = 0.5).csv"),
         get_assets_path())
     test_translation_dict = Dict("f_Hz"=>(1, Int64), "B_mT"=>(2, Float64), "FPS"=>(3, Float64))
@@ -23,4 +25,5 @@ end
     @test cd(()->isfile("test_animation.mp4"), get_assets_path())
     cd(()->rm("test_animation.mp4"), get_assets_path())
 
+    gr(show=true) # reenable showing plots
 end
