### 2023-11-09 9c6b078
try headless in CI.yml
gr(show=false) didn't work
