### 2023-11-09 c5c2bd8
add Project.toml to /paper
