+++ paper/Project.toml
@@ -1,0 +1,40 @@
+name = "MicroTracker"
+uuid = "f2f4b30c-0e7b-4018-b992-2f46df5706a6"
+authors = ["Coy Zimmermann"]
+version = "0.2.2"
+
+[deps]
+CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
+Conda = "8f4d0f93-b110-5947-807f-2305c1781a2d"
+DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
+DataFramesMeta = "1313f7d8-7da2-5740-9ea0-a2ca25f37964"
+FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
+FileIO = "5789e2e9-d7fb-5bc7-8068-2c6fae9b9549"
+ImageIO = "82e4d734-157c-48bb-816b-45c225c6df19"
+Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
+Optim = "429524aa-4258-5aef-a3af-852621145aeb"
+PyCall = "438e738f-606a-5dbb-bf0a-cddfbfd45ab0"
+Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
+Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
+StatsPlots = "f3b207a7-027a-5e70-b257-86293d7955fd"
+
+[compat]
+CSV = "0.10"
+Conda = "1"
+DataFrames = "1"
+DataFramesMeta = "0.14"
+FFTW = "1"
+FileIO = "1"
+ImageIO = "0.6"
+Optim = "1"
+PyCall = "1"
+Reexport = "1"
+Statistics = "1"
+StatsPlots = "0.15"
+julia = "1.7, 1.8, 1.9"
+
+[extras]
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["Test"]
