### 2023-11-09 fffa2ff
move $PYTHON setting into init
