### 2023-11-09 d0660ae
manually check if packages are installed
prevents a verbose Conda output everytime `using MicroTracker` is run
