### 2023-11-10 #5804 @czimm79
As promised, here is the current status on the review.

* The majority of code revisions are complete, and are ready for review in the three PRs referenced above.
* Manuscript revisions are nearing completion. I expect to share the new draft next week, along with responses to each reviewer comment regarding content.
* The transfer of the example videos from the package repository to the Julia Artifacts system is in progress. I'm currently working on making sure the artifact isn't downloaded multiple times, as the example videos are used both during the CI tests and the quick start guide. I expect this to be finished late next week.

Thank you all again for your patience and helpful comments.
