### 2023-12-18 #5804 @mkitti
My remaining comments concern the Python interface as detailed in https://github.com/czimm79/MicroTracker.jl/pull/47#pullrequestreview-1787821495 .

> It appears that PyCall currently [does not support](https://github.com/JuliaPy/Conda.jl/blob/master/README.md#conda-environments) non-root Conda environments

There is some misreading of the documentation here. `Conda.ROOTENV` is the default value for the optional `env` arguments across the Conda.jl API. However, you can specify the environment by passing a second argument to functions such as [`Conda.add`](https://github.com/JuliaPy/Conda.jl/blob/047bb989807c19a40c3fe68de996edc86133e107/src/Conda.jl#L326) and [`Conda.add_channel`](https://github.com/JuliaPy/Conda.jl/blob/047bb989807c19a40c3fe68de996edc86133e107/src/Conda.jl#L531).

PyCall.jl can be used with non-root environments by
1. Setting the environment variable `PYTHON` either via `ENV["PYTHON"]` in Julia or in the shell before starting Julia.
2. Calling `using Pkg; Pkg.build("PyCall")` as per the [PyCall.jl README](https://github.com/JuliaPy/PyCall.jl?tab=readme-ov-file#specifying-the-python-version).
