### 2024-01-08 b3721a0
getting segfault on local machine
adding Pkg as a dependency seems to have started testing failures. Using `Pkg.build("PyCall")` inside a precompile() safe module might be the issue.
