### 2024-01-10 f8e7732
only notify if ENV["PYTHON"] is set
