### 2024-01-12 e15080a
update docs to reflect env selection
