+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"settingupdev/#Setting-Up-(for-beginners)","page":"Setting Up (for beginners)","title":"Setting Up (for beginners)","text":"","category":"section"},{"location":"settingupdev/","page":"Setting Up (for beginners)","title":"Setting Up (for beginners)","text":"Using a notebook environment like Pluto or Jupyter doesn't work well when developing and contributing to an open source package. This page is a quick set up guide for using Visual Studio Code (VSCode) with the Julia extension to develop MicroTracker.jl.","category":"page"},{"location":"settingupdev/","page":"Setting Up (for beginners)","title":"Setting Up (for beginners)","text":"note: Note\nThis is by no means the only way to accomplish this. This is just the method I prefer and have found the easiest while developing MicroTracker.","category":"page"},{"location":"settingupdev/#VS-Code","page":"Setting Up (for beginners)","title":"VS Code","text":"","category":"section"},{"location":"settingupdev/","page":"Setting Up (for beginners)","title":"Setting Up (for beginners)","text":"Install Visual Studio Code and the Julia extension. \nEnsure you can run a hello world julia script, detailed in the getting started page.\nGet accustomed to running code in the integrated Julia REPL in VSCode using keybinds like shift+enter and ctrl+enter. This is also detailed in the extension docs here.\nAdd Revise.jl to your base Julia environment. By default, the VSCode Julia extension detects that its available and automatically loads it when starting a Julia extension integrated REPL in VSCode.\nInstall git.","category":"page"},{"location":"settingupdev/#GitHub-Desktop","page":"Setting Up (for beginners)","title":"GitHub Desktop","text":"","category":"section"},{"location":"settingupdev/","page":"Setting Up (for beginners)","title":"Setting Up (for beginners)","text":"Download Github Desktop.\nFile -> Clone Repository -> paste in the URL for MicroTracker.jl. https://github.com/czimm79/MicroTracker.jl.\nRight click on Current Repository (MicroTracker.jl) -> Open in Visual Studio Code.","category":"page"},{"location":"settingupdev/#Setting-up-environment","page":"Setting Up (for beginners)","title":"Setting up environment","text":"","category":"section"},{"location":"settingupdev/","page":"Setting Up (for beginners)","title":"Setting Up (for beginners)","text":"The dependent packages for MicroTracker normally automatically install behind the scenes when you use ] add MicroTracker in the REPL. When developing, we need to instantiate that dependency environment.","category":"page"},{"location":"settingupdev/","page":"Setting Up (for beginners)","title":"Setting Up (for beginners)","text":"Open the command palette in VSCode (ctrl+shift+p) and select Julia: Start REPL. Activate the MicroTracker environment:","category":"page"},{"location":"settingupdev/","page":"Setting Up (for beginners)","title":"Setting Up (for beginners)","text":"(v1.8) pkg> activate .","category":"page"},{"location":"settingupdev/","page":"Setting Up (for beginners)","title":"Setting Up (for beginners)","text":"You should be able to run the test command in the pkg mode and all tests should pass. You are now ready to make changes!","category":"page"},{"location":"settingupdev/","page":"Setting Up (for beginners)","title":"Setting Up (for beginners)","text":"note: Note\nWhen opening the package, sometimes VSCode will automatically recognize that the MicroTracker environment should be activated and prompt you. If this is the case, instead of (v1.8), you'll see MicroTracker, and you can skip step 2.","category":"page"},{"location":"settingupdev/#Creating-a-pull-request","page":"Setting Up (for beginner
[…]
