+++ Project.toml
@@ -32,7 +32,7 @@ PyCall = "1"
 Reexport = "1"
 Statistics = "1"
 StatsPlots = "0.15"
-julia = "1.7, 1.8, 1.9"
+julia = "1.7, 1.8, 1.9, 1.10"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
