### 2024-01-14 #5804 @czimm79
Hi @Kevin-Mattheus-Moerman - I believe we're ready to move forward. I've worked with Mark to address his comments over on https://github.com/czimm79/MicroTracker.jl/pull/47. The combined reviewer comments have been pushed to the master branch and tagged with v0.3. Let me know what I need to do next. Thanks!
