### 2024-01-17 #5804 @Kevin-Mattheus-Moerman
@mkitti @BioTurboNick @roflmaostc thanks for your excellent reviews! All boxes are now ticked. It looks like @BioTurboNick @roflmaostc have already given their blessing. @mkitti at this point would you recommend acceptance in JOSS as well? Thanks!
