### 2024-01-21 #5804 @Kevin-Mattheus-Moerman
@czimm79 it looks like all reviewers have now given their blessing. So we can proceed to process this work for acceptance in JOSS.
