### 2024-01-22 #5804 @czimm79
> # Post-Review Checklist for Editor and Authors
> ## Additional Author Tasks After Review is Complete
> * [ ]  Double check authors and affiliations (including ORCIDs)
> * [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.
> * [x]  Archive the release on Zenodo/figshare/etc and post the DOI here.
> * [ ]  Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.
> * [x]  Make sure that the license listed for the archive is the same as the software license.

Latest version: 0.3.2
Zenodo archive DOI: [![DOI](https://zenodo.org/badge/530280195.svg)](https://zenodo.org/doi/10.5281/zenodo.10547471)
