### 2024-01-28 #5804 @Kevin-Mattheus-Moerman
@czimm79 it looks like you have that reference twice in your .bib file. Can you check this please?
