### 2024-01-28 a1c4cad
remove duplicate bibtex
