+++ paper/paper.bib
@@ -225,16 +225,6 @@ Publisher: Nature Publishing Group},
 	pages = {80--90},
 }
 
-@misc{plas_fonspplutojl_2022,
-	title = {fonsp/{Pluto}.jl: v0.18.1},
-	url = {https://doi.org/10.5281/zenodo.6234110},
-	publisher = {Zenodo},
-	author = {Plas, Fons van der and Dral, Michiel and Berg, Paul and Γεωργακόπουλος, Παναγιώτης and Bochenski, Nicholas and Huijzer, Rik and Mengali, Alberto and Lungwitz, Benjamin and Burns, Connor and Ling, Jerry and Zhang, Eric and Schneider, Felipe S. S. and Weaver, Ian and {Rogerluo} and Kadowaki, Shuhei and Wu, Zihua and Gerritsen, Jelmar and Novosel, Rok and {Supanat} and Moon, Zachary and {pupuis} and Abbott, Michael and Bauer, Nicholas and Bouffard, Patrick and Terasaki, Satoshi and Polasa, Shashank and {TheCedarPrince} and {fghzxm}},
-	month = feb,
-	year = {2022},
-	doi = {10.5281/zenodo.6234110},
-}
-
 @article{wolvington_paddlebots_2023,
 	title = {Paddlebots: {Translation} of {Rotating} {Colloidal} {Assemblies} near an {Air}/{Water} {Interface}},
 	volume = {39},
