### 2024-01-29 #5804 @czimm79
Corrected. Released a new version v0.3.3. Does this work? Should I update the zenodo archive with the new version?
