### 2024-02-13 03819cf
update logo for dark mode
