### 2024-02-13 3f03ff4
add citation.cff file
